AST to IR
* improved the Bit type support
* added some new type generation utilities
* started with improved Derived implementation
* changed the AST interface

+++ src/Type.h
@@ -57,7 +57,7 @@ namespace libcasm_ir
 		typedef Rule*  RulePointer;
 		typedef u1     Boolean;
 		typedef i64    Integer;
-		typedef u8*    Bit;
+		typedef u64*   Bit;
 		typedef char*  String;
 		
 		enum ID
